TAB-6197 guard empty uni ids

+++ common/src/main/scala/uk/ac/warwick/tabula/data/MemberDao.scala
@@ -102,9 +102,11 @@ class MemberDaoImpl extends MemberDao with Logging with AttendanceMonitoringStud
 	}
 
 	def deleteByUniversityIds(universityIds: Seq[String]): Unit = {
-		val query = session.createQuery("delete from MEMBER where MEMBER.UNIVERSITYID in (:universityIds)")
-		query.setParameter("universityIds", universityIds.asJava)
-		query.executeUpdate()
+		if (universityIds.nonEmpty) {
+			val query = session.createQuery("delete from MEMBER where MEMBER.UNIVERSITYID in (:universityIds)")
+			query.setParameter("universityIds", universityIds.asJava)
+			query.executeUpdate()
+		}
 	}
 
 	def getByUniversityId(universityId: String, disableFilter: Boolean = false, eagerLoad: Boolean = false): Option[Member] = {
